feat: ✨ new subpackage roman_numeral
BREAKING CHANGE: 💥 roman_numeral script in miscellaneous_python subpackage replaced by a subpackage of the same name

The Roman decoding and encoding functionalities in the old script have grown so sophisticated that it can not be maintained as well as it should be. That's why it has been converted into a subpackage on its own.

Signed-off-by: Minh Tran <33189614+minhtrancccp@users.noreply.github.com>